Look for index.html in immediate parent of reveal.js.
It's nice to be able to use a common (custom) reveal.js library
for multiple presentations. One solution to this is to keep the
index.html file containing the presentation independent of the
reveal.js code itself. Have grunt look for index.html in the
immediate parent of the reveal.js checkout instead in the top-level
of the reveal.js checkout.